refac: product model분리, hook 연결
