+++ src/basic/models/cart.ts
@@ -38,16 +38,6 @@ export const calculateCartTotal = (
   };
 };
 
-// 상품의 남은 재고 계산
-export const getRemainingStock = (
-  product: Product,
-  cart: CartItem[]
-): number => {
-  const cartItem = cart.find((item) => item.product.id === product.id);
-  const remaining = product.stock - (cartItem?.quantity || 0);
-
-  return remaining;
-};
 
 // 장바구니에 상품 추가
 export const addItemToCart = (
